02/assignment/requirements.txt
# Assignment 02 Dependencies

# Core data processing
polars>=1.0.0
numpy>=1.24.0
pyyaml>=6.0

# Testing
pytest>=7.4.0

# Notebook support
jupyter>=1.0.0
jupytext>=1.15.0
nbconvert>=7.0.0
ipykernel>=6.25.0

# Data generation
faker>=20.0.0

# Optional
altair>=5.0.0
